各コントローラーのエラーハンドリング

+++ app/controllers/article_comments_controller.rb
@@ -2,13 +2,21 @@ class ArticleCommentsController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    article_comment = ArticleComment.create(article_comment_params)
-    redirect_to article_path(article_comment.article.id)
+    @article_comment = ArticleComment.new(article_comment_params)
+    if @article_comment.save
+      redirect_to article_path(@article_comment.article.id)
+    else
+      @article = Article.find(params[:article_id])
+      @article_comments  = @article.article_comments.includes(:user)
+      render "articles/show"
+    end
   end
   
   def destroy
     article_comment = ArticleComment.find(params[:id])
-    article_comment.destroy
+    if article_comment.user.id == current_user.id
+      article_comment.destroy
+    end
     redirect_to article_path(article_comment.article.id)
   end
